🏗 [portafly] moves data-list components into a shared folder and reduces imports versbosity
