+++ portafly/src/components/shared/data-list/index.ts
@@ -1,8 +1,9 @@
+export * from './modals'
+export * from './utils'
 export * from './BulkSelectorWidget'
 export * from './BulkActionsWidget'
 export * from './DataListContext'
 export * from './SearchWidget'
 export * from './Toolbar'
 export * from './PaginationWidget'
-export * from './modals'
-export * from './utils'
+export type { BulkAction } from './reducers'
