+++ portafly/src/components/shared/data-list/reducers/TableReducer.ts
