+++ portafly/src/components/shared/data-list/searchWidget.scss
