+++ portafly/src/tests/components/shared/data-list/DataListContext.test.tsx
@@ -1,7 +1,7 @@
-import { DataListProvider, useDataListFilters } from 'components/data-list'
-
 import React, { useState } from 'react'
+
 import { render, fireEvent } from '@testing-library/react'
+import { DataListProvider, useDataListFilters } from 'components'
 
 describe('DataListContext: useDataFilters', () => {
   const FilterComponent = () => {
