+++ portafly/src/tests/components/shared/data-list/modals/DataListModal.test.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
-import { DataListModal } from 'components/data-list/modals'
+
 import { render } from 'tests/custom-render'
 import { fireEvent } from '@testing-library/react'
+import { DataListModal } from 'components'
 
 it('shows up to 5 items and a button to expand the list', () => {
   const items = new Array(10).fill('test').map((admin, i) => admin + i)
